exclude port ranges when checking for port definition
Even though the port is technically "defined" when it's part of a range,
`semanage -m` exits with a "ValueError: Port _____ is not defined" when
trying to modify it. So the check for whether or not a port is defined
has to exclude port range definitions from its logic, as being part of a
range doesn't actually make the port "defined" for `semanage -m`
purposes.